+++ terraform/infra/modules/monitoring/alertmanager.tf
@@ -103,7 +103,7 @@ resource "aws_ecs_task_definition" "alertmanager" {
 
   container_definitions = jsonencode([{
     name              = "alertmanager"
-    image             = "${var.alertmanager_ecr_repository_url}:${var.image_tag}"
+    image             = local.service_image["alertmanager"]
     essential         = true
     memoryReservation = 128
     memory            = 256